+++ test/test.js
@@ -55,6 +55,11 @@ describe('openapi-format tests', () => {
             configFileOptions.sort = !configFileOptions['no-sort'];
             delete configFileOptions['no-sort'];
           }
+          configFileOptions.bundle = !configFileOptions['no-bundle'];
+          if (configFileOptions['no-sort'] && configFileOptions['no-bundle'] === true) {
+            configFileOptions.bundle = !configFileOptions['no-bundle'];
+            delete configFileOptions['no-bundle'];
+          }
           options = Object.assign({}, options, configFileOptions);
         }